Do not install/configure apache from this class
Configuring apache from within this module is a bad idea, since this
leads to conflicts. The module should assume that this is handled by an
apache module and only call apis.

Currently we do not have a module which provides an API, so also disable the
enablement of the conf.d file for now (must be handled in a site module
in the meanwhile)